Renamed to "schema-config-expose-admin-data"
