+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/blocks/schema-config-admin-fields/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "schema-config-admin-fields",
+	"name": "schema-config-expose-admin-data",
 	"version": "0.1.0",
 	"description": "Configure the 'Schema Expose Admin Data' module in the Schema Configuration",
 	"author": "Leonardo Losoviz",
