+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/blocks/schema-config-admin-fields/src/index.js
@@ -25,7 +25,7 @@ import {
  *
  * @see https://developer.wordpress.org/block-editor/developers/block-api/#registering-a-block
  */
-registerBlockType( 'graphql-api/schema-config-admin-fields', {
+registerBlockType( 'graphql-api/schema-config-expose-admin-data', {
 	/**
 	 * This is the display title for your block, which can be translated with `i18n` functions.
 	 * The block inserter will show this name.
