+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/blocks/schema-config-admin-fields/src/schema-config-expose-admin-data-card.js
